Fixes as per review Signed-off-by:Phillip Kruger <phillip.kruger@gmail.com>
