+++ server/implementation/src/main/java/io/smallrye/graphql/spi/datafetcher/CompletableFutureDataFetcherService.java
@@ -1,7 +1,5 @@
 package io.smallrye.graphql.spi.datafetcher;
 
-import static io.smallrye.graphql.SmallRyeGraphQLServerMessages.msg;
-
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionException;
@@ -12,14 +10,7 @@ import org.eclipse.microprofile.graphql.GraphQLException;
 
 import graphql.execution.DataFetcherResult;
 import graphql.schema.DataFetchingEnvironment;
-import io.smallrye.graphql.bootstrap.Config;
-import io.smallrye.graphql.execution.datafetcher.helper.ArgumentHelper;
-import io.smallrye.graphql.execution.datafetcher.helper.BatchLoaderHelper;
-import io.smallrye.graphql.execution.datafetcher.helper.FieldHelper;
-import io.smallrye.graphql.execution.datafetcher.helper.PartialResultHelper;
-import io.smallrye.graphql.execution.datafetcher.helper.ReflectionHelper;
-import io.smallrye.graphql.execution.event.EventEmitter;
-import io.smallrye.graphql.schema.model.Operation;
+import io.smallrye.graphql.SmallRyeGraphQLServerMessages;
 import io.smallrye.graphql.spi.DataFetcherService;
 import io.smallrye.graphql.transformation.AbstractDataFetcherException;
 
@@ -28,14 +19,7 @@ import io.smallrye.graphql.transformation.AbstractDataFetcherException;
  * 
  * @author Phillip Kruger (phillip.kruger@redhat.com)
  */
-public class CompletableFutureDataFetcherService implements DataFetcherService {
-    private Operation operation;
-    private FieldHelper fieldHelper;
-    private ReflectionHelper reflectionHelper;
-    private PartialResultHelper partialResultHelper;
-    private ArgumentHelper argumentHelper;
-    private EventEmitter eventEmitter;
-    private BatchLoaderHelper batchLoaderHelper;
+public class CompletableFutureDataFetcherService extends AbstractDataFetcherService {
 
     @Override
     public String getName() {
@@ -53,68 +37,41 @@ public class CompletableFutureDataFetcherService implements DataFetcherService {
     }
 
     @Override
-    public void init(Operation operation, Config config) {
-        this.operation = operation;
-        this.eventEmitter = EventEmitter.getInstance(config);
-        this.fieldHelper = new FieldHelper(operation);
-        this.reflectionHelper = new ReflectionHelper(operation, eventEmitter);
-        this.argumentHelper = new ArgumentHelper(operation.getArguments());
-        this.partialResultHelper = new PartialResultHelper();
-        this.batchLoaderHelper = new BatchLoaderHelper();
+    protected <T> T invokeAndTransform(DataFetchingEnvironment dfe, DataFetcherResult.Builder<Object> resultBuilder,
+            Object[] transformedArguments) throws AbstractDataFetcherException, Exception {
+        CompletionStage<Object> futureResultFromMethodCall = reflectionHelper.invoke(transformedArguments);
+
+        return (T) futureResultFromMethodCall.handle((result, throwable) -> {
+            if (throwable instanceof CompletionException) {
+                //Exception thrown by underlying method may be wrapped in CompletionException
+                throwable = throwable.getCause();
+            }
+
+            if (throwable != null) {
+                eventEmitter.fireOnDataFetchError(dfe.getExecutionId().toString(), throwable);
+                if (throwable instanceof GraphQLException) {
+                    GraphQLException graphQLException = (GraphQLException) throwable;
+                    partialResultHelper.appendPartialResult(resultBuilder, dfe, graphQLException);
+                } else if (throwable instanceof Exception) {
+                    throw SmallRyeGraphQLServerMessages.msg.dataFetcherException(operation, throwable);
+                } else if (throwable instanceof Error) {
+                    throw ((Error) throwable);
+                }
+            } else {
+                try {
+                    resultBuilder.data(fieldHelper.transformResponse(result));
+                } catch (AbstractDataFetcherException te) {
+                    te.appendDataFetcherResult(resultBuilder, dfe);
+                }
+            }
+
+            return resultBuilder.build();
+        });
     }
 
     @Override
-    public <T> T get(DataFetchingEnvironment dfe, DataFetcherResult.Builder<Object> resultBuilder) throws Exception {
-        eventEmitter.fireBeforeDataFetch();
-
-        try {
-            Object[] transformedArguments = argumentHelper.getArguments(dfe);
-            CompletionStage<Object> futureResultFromMethodCall = reflectionHelper.invoke(transformedArguments);
-
-            return (T) futureResultFromMethodCall.handle((result, throwable) -> {
-                if (throwable instanceof CompletionException) {
-                    //Exception thrown by underlying method may be wrapped in CompletionException
-                    throwable = throwable.getCause();
-                }
-
-                if (throwable != null) {
-                    eventEmitter.fireOnDataFetchError(dfe.getExecutionId().toString(), throwable);
-                    if (throwable instanceof GraphQLException) {
-                        GraphQLException graphQLException = (GraphQLException) throwable;
-                        partialResultHelper.appendPartialResult(resultBuilder, dfe, graphQLException);
-                    } else if (throwable instanceof Exception) {
-                        throw msg.dataFetcherException(operation, throwable);
-                    } else if (throwable instanceof Error) {
-                        throw ((Error) throwable);
-                    }
-                } else {
-                    try {
-                        resultBuilder.data(fieldHelper.transformResponse(result));
-                    } catch (AbstractDataFetcherException te) {
-                        te.appendDataFetcherResult(resultBuilder, dfe);
-                    }
-                }
-
-                return resultBuilder.build();
-            });
-
-        } catch (AbstractDataFetcherException abstractDataFetcherException) {
-            //Arguments or result couldn't be transformed
-            abstractDataFetcherException.appendDataFetcherResult(resultBuilder, dfe);
-            eventEmitter.fireOnDataFetchError(dfe.getExecutionId().toString(), abstractDataFetcherException);
-        } catch (GraphQLException graphQLException) {
-            partialResultHelper.appendPartialResult(resultBuilder, dfe, graphQLException);
-            eventEmitter.fireOnDataFetchError(dfe.getExecutionId().toString(), graphQLException);
-        } catch (SecurityException | IllegalAccessException | IllegalArgumentException ex) {
-            //m.invoke failed
-            eventEmitter.fireOnDataFetchError(dfe.getExecutionId().toString(), ex);
-            throw ex;
-        } finally {
-            eventEmitter.fireAfterDataFetch();
-        }
-
+    protected <T> T invokeFailure(DataFetcherResult.Builder<Object> resultBuilder) {
         return (T) CompletableFuture.completedFuture(resultBuilder.build());
-
     }
 
     @Override
@@ -123,4 +80,5 @@ public class CompletableFutureDataFetcherService implements DataFetcherService {
         final ClassLoader tccl = Thread.currentThread().getContextClassLoader();
         return (CompletableFuture<List<T>>) reflectionHelper.invokePrivileged(tccl, arguments);
     }
+
 }
