--- server/implementation/src/main/resources/META-INF/services/io.smallrye.graphql.spi.DataFetcherService
@@ -1,1 +1,0 @@
-io.smallrye.graphql.spi.datafetcher.CompletableFutureDataFetcherService
